fix issues when there's no APD to load, etc
