+++ web/src/reducers/apd.js
@@ -1,5 +1,6 @@
 import u from 'updeep';
 
+import { defaultAPDYearOptions, defaultAPDYears } from '../util';
 import {
   GET_APD_REQUEST,
   GET_APD_SUCCESS,
@@ -7,29 +8,11 @@ import {
   UPDATE_APD
 } from '../actions/apd';
 
-const thisFFY = (() => {
-  const year = new Date().getFullYear();
-
-  // Federal fiscal year starts October 1,
-  // but Javascript months start with 0 for
-  // some reason, so October is month 9.
-  if (new Date().getMonth() > 8) {
-    return year + 1;
-  }
-  return year;
-})();
-
-// The UI turns the years into strings, so let's
-// just make them strings in the state as well;
-// that simplifies things
-const threeYears = [thisFFY, thisFFY + 1, thisFFY + 2].map(y => `${y}`);
-const firstTwoYears = threeYears.slice(0, 2);
-
 const initialState = {
   data: {
     id: '',
-    years: firstTwoYears,
-    yearOptions: threeYears,
+    years: defaultAPDYears,
+    yearOptions: defaultAPDYearOptions,
     overview: '',
     hitNarrative: '',
     hieNarrative: '',
@@ -54,7 +37,8 @@ const reducer = (state = initialState, action) => {
         narrativeHIE: hieNarrative,
         narrativeMMIS: mmisNarrative,
         previousActivitySummary
-      } = action.data;
+      } =
+        action.data || {};
 
       return {
         ...state,
@@ -66,8 +50,8 @@ const reducer = (state = initialState, action) => {
           hitNarrative,
           hieNarrative,
           mmisNarrative,
-          years: years.map(y => `${y}`) || firstTwoYears,
-          yearOptions: threeYears,
+          years: (years || defaultAPDYears).map(y => `${y}`),
+          yearOptions: defaultAPDYearOptions,
           previousActivitySummary: previousActivitySummary || ''
         }
       };
